+++ cypress/pageObjects/JobConfigurePage.js
@@ -1,19 +1,11 @@
-import JobProfilePage from "../pageObjects/JobProfilePage";
-import HomePage from "../pageObjects/HomePage";
+import JobProfilePage from "./MultiConfigurationProjectConfigurePage";
 
 class JobConfigurePage {
   getSaveButton = () => cy.get("button[name='Submit']");
-  getDeleteLink = () => cy.get('a[data-message^="Delete"]');
 
   clickSaveButton() {
     this.getSaveButton().click();
     return new JobProfilePage();
   }
-
-  clickDeleteLink() {
-    this.getDeleteLink().click();
-    return new HomePage();
-  }
 }
-
 export default JobConfigurePage;
