--- cypress/pageObjects/JobProfilePage.js
@@ -1,12 +1,0 @@
-import HomePage from "../pageObjects/HomePage";
-import JobConfigurePage from "../pageObjects/JobConfigurePage";
-
-class JobProfilePage {
-  getDeleteLink = () => cy.get('a[data-message^="Delete"]');
-
-  clickDeleteLink() {
-    this.getDeleteLink().click();
-    return new HomePage();
-  }
-}
-export default JobProfilePage;
